chg: refactor: move spoof to file
